myapp/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.8
exceptiongroup==1.2.0
gunicorn==21.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg==3.1.14
psycopg-binary==3.1.14
pycodestyle==2.11.1
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
whitenoise==6.6.0
ollama==0.1.5
